- moving bcond.hpp to detail/bcond_common.hpp (and renaming the class) - moving cyclic_common.hpp and polar_common.hpp to detail (+namespace) - removing open_common.file - was not use anywhere - moving shared bc to a separate file - removing Doxygen headers
