+++ libmpdata++/solvers/detail/solver_common.hpp
@@ -12,7 +12,7 @@
 
 #include <libmpdata++/solvers/detail/monitor.hpp>
 
-#include <libmpdata++/bcond/bcond.hpp>
+#include <libmpdata++/bcond/detail/bcond_common.hpp>
 
 #include <array>
 
@@ -41,7 +41,7 @@ namespace libmpdataxx
 
         typedef typename ct_params_t::real_t real_t;
         typedef blitz::Array<real_t, n_dims> arr_t;
-        using bcp_t = std::unique_ptr<bcond::bcond_t<real_t>>;
+        using bcp_t = std::unique_ptr<bcond::detail::bcond_common<real_t>>;
 
         using ix = typename ct_params_t::ix;
